Apply review conventions to transport-assignment package
- Rename package from movici-sim-transport-assignment to movici-transport-assignment-model
- Flatten test directory: move tests directly under tests/ with unique names
- Merge models/conftest.py fixtures into root conftest.py
- Remove ruff config from package pyproject.toml, consolidate in root
- Add README.rst for PyPI publishing
- Add per-package Makefile mirroring simulation-core
- Remove tests/__init__.py to avoid conftest collision when running all tests
