+++ packages/transport-assignment/src/movici_transport_assignment_model/integrations/collections.py
