+++ packages/transport-assignment/src/movici_transport_assignment_model/json_schemas/__init__.py
