+++ packages/transport-assignment/src/movici_transport_assignment_model/json_schemas/legacy/traffic_assignment_calculation.json
