+++ packages/transport-assignment/src/movici_transport_assignment_model/json_schemas/traffic_assignment_calculation.json
