+++ packages/transport-assignment/tests/test_ae_point_generator.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from movici_sim_transport_assignment.integrations.point_generator import PointGenerator
+from movici_transport_assignment_model.integrations.point_generator import PointGenerator
 
 
 def test_can_add_points():
